requirements.txt
playwright
openai-whisper
sounddevice
soundfile
numpy
jinja2
weasyprint
nest_asyncio
